feat: add diff adaption
Adds a new adaption that allows the user to select the files they want to include in the diff, if the generated diff is too long. The selected files are then used to generate the diff. The diff adaption also reads the diff from a file called "diff.txt" and returns the diff and the number of tokens used.